+++ GliaWidgets/Sources/Component/ImageView/File/FilePreviewView.Environment.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+extension FilePreviewView {
+    struct Environment {
+        var uiScreen: UIKitBased.UIScreen
+    }
+}
+
+extension FilePreviewView.Environment {
+//    static func create(with environment: SecureConversations.FileUploadView.Environment) -> Self {
+//        .init(uiScreen: environment.uiScreen)
+//    }
+}
