+++ GliaWidgets/Sources/ViewController/Common/FilePicker/FilePickerController.Envrionment.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+extension FilePickerController {
+    public struct Environment {
+        let fileManager: FoundationBased.FileManager
+    }
+}
+
+extension FilePickerController.Environment {
+    static func create(with environment: SecureConversations.Coordinator.Environment) -> Self {
+        .init(fileManager: environment.fileManager)
+    }
+}
