+++ GliaWidgets/Sources/ViewModel/Common/FilePicker/FilePickerViewModel.Environment.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+extension FilePickerViewModel {
+    struct Environment {
+        var log: CoreSdkClient.Logger
+    }
+}
+
+extension FilePickerViewModel.Environment {
+    static func create(with environment: SecureConversations.Coordinator.Environment) -> Self {
+        .init(log: environment.log)
+    }
+}
