+++ GliaWidgets/Sources/ViewModel/Common/FilePicker/FilePickerViewModel.swift
@@ -68,9 +68,3 @@ class FilePickerViewModel: ViewModel {
         }
     }
 }
-
-extension FilePickerViewModel {
-    struct Environment {
-        var log: CoreSdkClient.Logger
-    }
-}
